Updated giveKey and givePouch
Modified the key and pouch giving systems to allow keys and pouches to be automatically stacked, instead of giving one per empty slot. Remaining items are automatically dropped in the player's location. It detects both empty slots and partially full items of the ItemStack, depending on its maximum stack size, to set the maximum possible amount of keys/pouches in the player's inventory before dropping the rest.